32-bit vs. 64-bit int/32-bit addr vs. 64-bit int/64-bit addr
Discovered why the VAX wasn't working: the simhcore library has to match
the other compile flags, principally the USE_INT64 and USE_ADDR64. The
makefile gets away with this because the simulator source files are
always compiled with the respective simulator. When we use libraries,
need to match the defines sent to the compiler. Otherwise, weird bugs
ensue.