cmd_flag_t: Command flag argument
Command function's first (or fourth) argument is int32 and used to keep track of script recursion depth. size_t really should be used for indexing arrays, which is generally what the command flag is used for. First step: change the type to something meaningful and deal with all of the special cases later.