Rethink NO_SYSTEM_ENVIRONMENT_PATH
Found why FindPCRE2 didn't find pcre2 on MinGW: It needed better names
for the libraries being searched (i.e., libpcre2 and pcre2). This is
somewhat academic if the MinGW developer installed pkgconfig, CMake
finds pkgconfig and has to use it as a fallback to the usual
find_path/find_library sequence in the Find* modules.

CMake's find_library will still find debug versions of libraries if
they're installed in the simh/BIN directory, which is independent of the
NO_SYSTEM_ENVIRONMENT_PATH.