brotlidec, harfbuzz link dependencies for freetype2
Freetype2 detects the Google Brotli compression decoder and Harfbuzz
libraries when building as a dependency. SIMH won't build Brotli or
Harfbuzz as additional dependencies, but SIMH does have to provide the
include and library paths to ensure no unresolved symbols.

Shows up as a link error in the BESM6 simulator.